Fix multiple bugs in Blend.
There were many cases were the value should be clipped if greater than 255.
The previous code would just wrap-around. Note that this is for the default
C++ path that's never taken if SIMD is available. To test this, I disabled
the detection of SIMD instructions.

Fixed a bug in the SIMD code that compared the end of the table / 2, rather
than just the end of the table. Most of the last half of the tables have no
implementation but a few do. Now MULTIPLY and the other are excuted using the
SIMD operation. I verified that the new warning was no longer triggered and
that the code passed validation.

Finally, when using LaunchOptions, the out ptr is already increment for the x
value where we supposed to start. That's done by the code that walks in 2D,
before the kernel is called. The SIMD code added the starting x once again.
This went undetected because there were no tests case for passing LaunchOptions
with real values.

Bug: 178718752
Test: Ran the RS Blend test in ImageProcessing & other tests.
Change-Id: I060da5ed704d0ce589208b7ae021fca93644cc6f
Merged-In: I060da5ed704d0ce589208b7ae021fca93644cc6f
(cherry picked from commit 2f7c3e259eff09bcc3e44c7a11c0fbfaa1858349)