Prevent IntrinsicResize from adding xstart twice when using LaunchOptions.
The cpu_ref implementation of IntrinsicResize has a bug when used with LaunchOptions.

walk_2d_foreach in rsCpuCore calls FepPtrSetup. This sets the outPtr[0] to point to the first byte that needs to be filled. It takes into account the starting x value provided in the LaunchOptions.

The cpu_ref code for resize adds the xstart value yet again before storing bytes, placing results at the incorrect location.

The bug was not previously found because this was not tested.

Bug: 180180223
Test: Modified the IntrinsicResize test and ran it.
Change-Id: I0cb5414834d6c8204fa0a266f5e4d7e9c77f5c71