Copy the intrinsic files we'll reuse for the Toolkit.
Copy and rename the various CPU implementations of the RenderScript
Intrinsics found in frameworks/rs/cpu_ref that we'll reuse for the Toolkit.

Besides copying and renaming the files, a few files were modified to
pass the lint checker. Only changes were avoiding going over 100 characters
and fixing some typo in the comments.

This is the first of many successive CLs. The end goal is
to publish this on github or through JetPack.

Bug: 178476084
Test: Compiled. No tests yet.
Change-Id: I764aba9bf21e24c2c81ba97a56bf9bdf79b30d63