Reland: "Switch RS runtime libraries to prebuilt build rules"
This change also deletes the Android.mk files for building the runtimes,
but doesn't delete the source files just yet. That can be done in a
follow-up, once we are sure that the prebuilts are sufficient.

This relands I621e9be6ffdae4b9e03d33aeae9f0f82b0675060 after fixing
the prebuilt_renderscript_bitcode rules to avoid install conflicts 
in NDK builds.

Bug: http://b/301495450
Bug: http://b/307386007
Change-Id: I2f0f96ab1b739e9401d454f64c4ad768fb7963a4
Test: cd cts/tests/tests/renderscript/src/android/renderscript/cts && atest .